Renamed Dump20190814.sql to Database.sql
This file, currently gives you all the missing data and tables you need to Develop MS websites along with the necessary Data.

When the need arise, we shall add more data whenever it's required for out stage of Development and the features we are working on.

Reveals one of my many alt emails, you can still access the removed file for reference, doesn't matter.